refactoring and code splitting
